Fixed function signatures

+++ utilities/utilities.c
@@ -152,12 +152,12 @@ void dev_null(int count, ...)
     (void)(count);
 }
 
-int max(const double x, const double y)
+double max(const double x, const double y)
 {
     return (x > y) ? x : y;
 }
 
-int min(const double x, const double y)
+double min(const double x, const double y)
 {
     return (x < y) ? x : y;
 }
